requirements.txt
Django==4.2.5
newsapi-python==0.2.7
requests==2.32.3
numpy==1.24.4
Pillow==9.5.0
imageio==2.31.1
django-tinymce==4.1.0
Whoosh==2.7.4
django-haystack==3.3.0
django-clearcache==1.2.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-simple-history==3.7.0